+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning].
 
 - RandomSkunk.Results:
     - Allow Error properties to be initialized with null values. Without this change, there was no way to use a `with` statement to create a new Error with a null `Identifier` or `StackTrace` value.
+    - Adds `isSensitive` and `extensions` parameters to result `Fail` methods.
 
 ## [1.0.0-alpha22] - 2022-12-09